+++ react/features/dynamic-branding/actions.js
@@ -9,7 +9,7 @@ import {
     SET_DYNAMIC_BRANDING_FAILED,
     SET_DYNAMIC_BRANDING_READY
 } from './actionTypes';
-import { getDynamicBrandingUrl } from './functions';
+import { getDynamicBrandingUrl } from './functions.any';
 
 const logger = getLogger(__filename);
